+++ src/main/java/net/geforcemods/securitycraft/blockentities/RiftStabilizerBlockEntity.java
@@ -1,0 +1,365 @@
+package net.geforcemods.securitycraft.blockentities;
+
+import java.util.Arrays;
+import java.util.EnumMap;
+import java.util.Map;
+import java.util.function.ToIntFunction;
+
+import net.geforcemods.securitycraft.ClientHandler;
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
+import net.geforcemods.securitycraft.api.Option;
+import net.geforcemods.securitycraft.api.Option.BooleanOption;
+import net.geforcemods.securitycraft.api.Option.DisabledOption;
+import net.geforcemods.securitycraft.api.Option.IntOption;
+import net.geforcemods.securitycraft.blockentities.RiftStabilizerBlockEntity.TeleportationType;
+import net.geforcemods.securitycraft.blocks.RiftStabilizerBlock;
+import net.geforcemods.securitycraft.misc.BlockEntityTracker;
+import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.network.server.SyncRiftStabilizer;
+import net.geforcemods.securitycraft.util.ITickingBlockEntity;
+import net.geforcemods.securitycraft.util.IToggleableEntries;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.Tag;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.monster.EnderMan;
+import net.minecraft.world.entity.monster.Shulker;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.event.entity.EntityTeleportEvent;
+import net.minecraftforge.event.entity.EntityTeleportEvent.ChorusFruit;
+import net.minecraftforge.event.entity.EntityTeleportEvent.EnderEntity;
+import net.minecraftforge.event.entity.EntityTeleportEvent.EnderPearl;
+import net.minecraftforge.event.entity.EntityTeleportEvent.SpreadPlayersCommand;
+import net.minecraftforge.event.entity.EntityTeleportEvent.TeleportCommand;
+
+public class RiftStabilizerBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity, IToggleableEntries<TeleportationType> {
+	private final IntOption signalLength = new IntOption(this::getBlockPos, "signalLength", 60, 5, 400, 5, true); //20 seconds max
+	private final IntOption range = new IntOption(this::getBlockPos, "range", 5, 1, 15, 1, true);
+	private final DisabledOption disabled = new DisabledOption(false);
+	private final Map<TeleportationType, Boolean> teleportationFilter = new EnumMap<>(TeleportationType.class);
+	private double lastTeleportDistance;
+	private TeleportationType lastTeleportationType;
+	private boolean tracked = false;
+
+	public RiftStabilizerBlockEntity(BlockPos pos, BlockState state) {
+		super(SCContent.RIFT_STABILIZER_BLOCK_ENTITY.get(), pos, state);
+		//when adding new types ONLY ADD TO THE END. anything else will break saved data.
+		//ordering is done in RiftStabilizerScreen based on the user's current language
+		teleportationFilter.put(TeleportationType.CHORUS_FRUIT, true);
+		teleportationFilter.put(TeleportationType.ENDER_PEARL, true);
+		teleportationFilter.put(TeleportationType.ENDERMAN, false);
+		teleportationFilter.put(TeleportationType.SHULKER, false);
+		teleportationFilter.put(TeleportationType.MODDED, false);
+	}
+
+	@Override
+	public void tick(Level level, BlockPos pos, BlockState state) {
+		if (!tracked) {
+			BlockEntityTracker.RIFT_STABILIZER.track(this);
+			tracked = true;
+		}
+	}
+
+	@Override
+	public void setRemoved() {
+		super.setRemoved();
+
+		// Stop tracking Rift Stabilizers when they are removed from the world
+		BlockEntityTracker.RIFT_STABILIZER.stopTracking(this);
+	}
+
+	@Override
+	public void setFilter(TeleportationType teleportationType, boolean allowed) {
+		if (teleportationFilter.containsKey(teleportationType)) {
+			teleportationFilter.put(teleportationType, allowed);
+			setChanged();
+
+			if (level.isClientSide)
+				SecurityCraft.channel.sendToServer(new SyncRiftStabilizer(worldPosition, teleportationType, allowed));
+
+			RiftStabilizerBlockEntity connectedBlockEntity = RiftStabilizerBlock.getConnectedBlockEntity(level, worldPosition);
+
+			if (connectedBlockEntity != null) {
+				connectedBlockEntity.teleportationFilter.put(teleportationType, allowed);
+				connectedBlockEntity.setChanged();
+
+				if (level.isClientSide)
+					SecurityCraft.channel.sendToServer(new SyncRiftStabilizer(connectedBlockEntity.worldPosition, teleportationType, allowed));
+			}
+		}
+	}
+
+	@Override
+	public boolean getFilter(TeleportationType teleportationType) {
+		return teleportationFilter.containsKey(teleportationType) && teleportationFilter.get(teleportationType);
+	}
+
+	@Override
+	public ToIntFunction<TeleportationType> getComparatorOutputFunction() {
+		return t -> Arrays.asList(TeleportationType.values()).indexOf(t) + 1;
+	}
+
+	@Override
+	public Map<TeleportationType, Boolean> getFilters() {
+		return teleportationFilter;
+	}
+
+	@Override
+	public TeleportationType getDefaultType() {
+		return TeleportationType.MODDED;
+	}
+
+	@Override
+	public CompoundTag save(CompoundTag tag) {
+		super.save(tag);
+
+		CompoundTag teleportationNBT = new CompoundTag();
+		int i = 0;
+
+		for (boolean b : teleportationFilter.values()) {
+			teleportationNBT.putBoolean("teleportationType" + i, b);
+			i++;
+		}
+
+		tag.put("teleportationTypes", teleportationNBT);
+		tag.putDouble("lastTeleportDistance", lastTeleportDistance);
+
+		if (lastTeleportationType != null)
+			tag.putInt("lastTeleportationType", Arrays.asList(TeleportationType.values()).indexOf(lastTeleportationType));
+
+		return tag;
+	}
+
+	@Override
+	public void load(CompoundTag tag) {
+		super.load(tag);
+
+		if (tag.contains("teleportationTypes", Tag.TAG_COMPOUND)) {
+			CompoundTag teleportationNBT = tag.getCompound("teleportationTypes");
+			int i = 0;
+
+			for (TeleportationType teleportationType : teleportationFilter.keySet()) {
+				teleportationFilter.put(teleportationType, teleportationNBT.getBoolean("teleportationType" + i));
+				i++;
+			}
+		}
+
+		lastTeleportDistance = tag.getDouble("lastTeleportDistance");
+
+		if (tag.contains("lastTeleportationType"))
+			lastTeleportationType = TeleportationType.values()[tag.getInt("lastTeleportationType")];
+	}
+
+	@Override
+	public void onModuleInserted(ItemStack stack, ModuleType module, boolean toggled) {
+		super.onModuleInserted(stack, module, toggled);
+
+		RiftStabilizerBlockEntity connectedBlockEntity = RiftStabilizerBlock.getConnectedBlockEntity(level, worldPosition);
+
+		if (connectedBlockEntity != null) {
+			if (toggled ? !connectedBlockEntity.isModuleEnabled(module) : !connectedBlockEntity.hasModule(module))
+				connectedBlockEntity.insertModule(stack, toggled);
+		}
+
+		if (module == ModuleType.DISGUISE) {
+			onInsertDisguiseModule(this, stack);
+
+			if (connectedBlockEntity != null)
+				onInsertDisguiseModule(connectedBlockEntity, stack);
+		}
+	}
+
+	@Override
+	public void onModuleRemoved(ItemStack stack, ModuleType module, boolean toggled) {
+		super.onModuleRemoved(stack, module, toggled);
+
+		RiftStabilizerBlockEntity connectedBlockEntity = RiftStabilizerBlock.getConnectedBlockEntity(level, worldPosition);
+
+		if (connectedBlockEntity != null) {
+			if (toggled ? connectedBlockEntity.isModuleEnabled(module) : connectedBlockEntity.hasModule(module))
+				connectedBlockEntity.removeModule(module, toggled);
+		}
+
+		if (module == ModuleType.DISGUISE) {
+			onRemoveDisguiseModule(this, stack);
+
+			if (connectedBlockEntity != null)
+				onRemoveDisguiseModule(connectedBlockEntity, stack);
+		}
+		else if (module == ModuleType.SMART) {
+			onRemoveSmartModule(this);
+
+			if (connectedBlockEntity != null)
+				onRemoveSmartModule(connectedBlockEntity);
+		}
+	}
+
+	private void onInsertDisguiseModule(BlockEntity be, ItemStack stack) {
+		if (!be.getLevel().isClientSide)
+			be.getLevel().sendBlockUpdated(be.getBlockPos(), be.getBlockState(), be.getBlockState(), 3);
+		else
+			ClientHandler.putDisguisedBeRenderer(be, stack);
+	}
+
+	private void onRemoveDisguiseModule(BlockEntity be, ItemStack stack) {
+		if (!be.getLevel().isClientSide)
+			be.getLevel().sendBlockUpdated(be.getBlockPos(), be.getBlockState(), be.getBlockState(), 3);
+		else
+			ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(be);
+	}
+
+	private void onRemoveSmartModule(RiftStabilizerBlockEntity be) {
+		be.teleportationFilter.put(TeleportationType.CHORUS_FRUIT, true);
+		be.teleportationFilter.put(TeleportationType.ENDER_PEARL, true);
+		be.teleportationFilter.put(TeleportationType.ENDERMAN, false);
+		be.teleportationFilter.put(TeleportationType.SHULKER, false);
+		be.teleportationFilter.put(TeleportationType.MODDED, false);
+	}
+
+	public void setLastTeleport(double teleportDistance, TeleportationType type) {
+		lastTeleportDistance = teleportDistance;
+		lastTeleportationType = type;
+	}
+
+	public double getLastTeleportDistance() {
+		return lastTeleportDistance;
+	}
+
+	public TeleportationType getLastTeleportationType() {
+		return lastTeleportationType;
+	}
+
+	@Override
+	public ModuleType[] acceptedModules() {
+		return new ModuleType[] {
+				ModuleType.ALLOWLIST, ModuleType.DISGUISE, ModuleType.REDSTONE, ModuleType.HARMING, ModuleType.SMART
+		};
+	}
+
+	@Override
+	public void onOptionChanged(Option<?> option) {
+		RiftStabilizerBlockEntity connectedBlockEntity = RiftStabilizerBlock.getConnectedBlockEntity(level, worldPosition);
+
+		if (option.getName().equals("signalLength")) {
+			IntOption intOption = (IntOption) option;
+
+			if (connectedBlockEntity != null)
+				connectedBlockEntity.setSignalLength(intOption.get());
+		}
+		else if (option.getName().equals("range")) {
+			IntOption intOption = (IntOption) option;
+
+			if (connectedBlockEntity != null)
+				connectedBlockEntity.setRange(intOption.get());
+		}
+		else if (option.getName().equals("disabled")) {
+			BooleanOption bo = (BooleanOption) option;
+
+			if (connectedBlockEntity != null)
+				connectedBlockEntity.setDisabled(bo.get());
+		}
+
+		super.onOptionChanged(option);
+	}
+
+	@Override
+	public Option<?>[] customOptions() {
+		return new Option[] {
+				signalLength, range, disabled
+		};
+	}
+
+	@Override
+	public void onOwnerChanged(BlockState state, Level level, BlockPos pos, Player player) {
+		RiftStabilizerBlockEntity be = RiftStabilizerBlock.getConnectedBlockEntity(level, pos);
+
+		if (be != null) {
+			be.setOwner(getOwner().getUUID(), getOwner().getName());
+
+			if (!level.isClientSide)
+				level.getServer().getPlayerList().broadcastAll(be.getUpdatePacket());
+		}
+
+		super.onOwnerChanged(state, level, pos, player);
+	}
+
+	public void setSignalLength(int signalLength) {
+		if (getSignalLength() != signalLength) {
+			this.signalLength.setValue(signalLength);
+			level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 3); //sync option change to client
+			setChanged();
+		}
+	}
+
+	public int getSignalLength() {
+		return signalLength.get();
+	}
+
+	public void setRange(int range) {
+		if (getRange() != range) {
+			this.range.setValue(range);
+			level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 3); //sync option change to client
+			setChanged();
+		}
+	}
+
+	public int getRange() {
+		return range.get();
+	}
+
+	public void setDisabled(boolean disabled) {
+		if (isDisabled() != disabled) {
+			this.disabled.setValue(disabled);
+			level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 3); //sync option change to client
+			setChanged();
+		}
+	}
+
+	public boolean isDisabled() {
+		return disabled.get();
+	}
+
+	public enum TeleportationType {
+		CHORUS_FRUIT(Items.CHORUS_FRUIT.getDescriptionId()),
+		ENDER_PEARL(Items.ENDER_PEARL.getDescriptionId()),
+		ENDERMAN(EntityType.ENDERMAN.getDescriptionId()),
+		SHULKER(EntityType.SHULKER.getDescriptionId()),
+		MODDED("gui.securitycraft:rift_stabilizer.modded");
+
+		public final String label;
+
+		TeleportationType(String label) {
+			this.label = label;
+		}
+
+		public static TeleportationType getTypeFromEvent(EntityTeleportEvent event) {
+			if (event instanceof ChorusFruit)
+				return CHORUS_FRUIT;
+			else if (event instanceof EnderPearl)
+				return ENDER_PEARL;
+			else if (event instanceof EnderEntity enderEntityEvent) {
+				if (enderEntityEvent.getEntityLiving() instanceof EnderMan)
+					return ENDERMAN;
+				else if (enderEntityEvent.getEntityLiving() instanceof Shulker)
+					return SHULKER;
+
+				return MODDED;
+			}
+			else if (event instanceof TeleportCommand || event instanceof SpreadPlayersCommand)
+				return null;
+
+			return MODDED;
+		}
+
+		@Override
+		public String toString() {
+			return label;
+		}
+	}
+}