+++ src/main/java/net/geforcemods/securitycraft/blockentities/TrophySystemBlockEntity.java
@@ -21,6 +21,7 @@ import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.network.client.SetTrophySystemTarget;
 import net.geforcemods.securitycraft.network.server.SyncTrophySystem;
 import net.geforcemods.securitycraft.util.ITickingBlockEntity;
+import net.geforcemods.securitycraft.util.IToggleableEntries;
 import net.geforcemods.securitycraft.util.ModuleUtils;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.minecraft.core.BlockPos;
@@ -41,7 +42,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraftforge.fmllegacy.network.PacketDistributor;
 
-public class TrophySystemBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity, ILockable {
+public class TrophySystemBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity, ILockable, IToggleableEntries<EntityType<?>> {
 	/* The range (in blocks) that the trophy system will search for projectiles in */
 	public static final int RANGE = 10;
 	/*
@@ -221,10 +222,7 @@ public class TrophySystemBlockEntity extends DisguisableBlockEntity implements I
 		return owner == null || (!owner.owns(this) && !ModuleUtils.isAllowed(this, owner.getName()));
 	}
 
-	public void toggleFilter(EntityType<?> projectileType) {
-		setFilter(projectileType, !projectileFilter.get(projectileType));
-	}
-
+	@Override
 	public void setFilter(EntityType<?> projectileType, boolean allowed) {
 		if (projectileFilter.containsKey(projectileType)) {
 			projectileFilter.put(projectileType, allowed);
@@ -235,12 +233,24 @@ public class TrophySystemBlockEntity extends DisguisableBlockEntity implements I
 		}
 	}
 
+	@Override
 	public boolean getFilter(EntityType<?> projectileType) {
 		return projectileFilter.get(projectileType);
 	}
 
+	@Override
 	public Map<EntityType<?>, Boolean> getFilters() {
 		return projectileFilter;
+	}
+
+	@Override
+	public EntityType<?> getDefaultType() {
+		return EntityType.PIG;
+	}
+
+	@Override
+	public String getDefaultTypeName() {
+		return "gui.securitycraft:trophy_system.moddedProjectiles";
 	}
 
 	public boolean isDisabled() {