+++ src/main/java/net/geforcemods/securitycraft/screen/ToggleListScreen.java
@@ -1,0 +1,206 @@
+package net.geforcemods.securitycraft.screen;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.BufferBuilder;
+import com.mojang.blaze3d.vertex.BufferUploader;
+import com.mojang.blaze3d.vertex.DefaultVertexFormat;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.Tesselator;
+import com.mojang.blaze3d.vertex.VertexFormat.Mode;
+
+import net.geforcemods.securitycraft.SecurityCraft;
+import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.util.ClientUtils;
+import net.geforcemods.securitycraft.util.IToggleableEntries;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.resources.sounds.SimpleSoundInstance;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraftforge.client.gui.ScrollPanel;
+
+public class ToggleListScreen<T> extends Screen {
+	private static final ResourceLocation GUI_TEXTURE = new ResourceLocation(SecurityCraft.MODID, "textures/gui/container/blank.png");
+	public final Component scrollListTitle, moduleRequired, toggle;
+	private final int imageWidth = 176;
+	private final int imageHeight = 166;
+	private int leftPos;
+	private int topPos;
+	private final boolean isSmart;
+	private final boolean isRedstone;
+	private final IToggleableEntries<T> be;
+	private ToggleScrollList teleportationModeList;
+
+	public ToggleListScreen(IToggleableEntries<T> be, String title, Component scrollListTitle, Component moduleRequired, Component toggle) {
+		super(new TranslatableComponent(title));
+
+		this.be = be;
+		isSmart = be instanceof IModuleInventory moduleInventory && moduleInventory.isModuleEnabled(ModuleType.SMART);
+		isRedstone = be instanceof IModuleInventory moduleInventory && moduleInventory.isModuleEnabled(ModuleType.REDSTONE);
+		this.scrollListTitle = scrollListTitle;
+		this.moduleRequired = moduleRequired;
+		this.toggle = toggle;
+	}
+
+	@Override
+	protected void init() {
+		super.init();
+
+		leftPos = (width - imageWidth) / 2;
+		topPos = (height - imageHeight) / 2;
+		addRenderableWidget(teleportationModeList = new ToggleScrollList(minecraft, imageWidth - 24, imageHeight - 60, topPos + 40, leftPos + 12));
+	}
+
+	@Override
+	public void render(PoseStack pose, int mouseX, int mouseY, float partialTicks) {
+		renderBackground(pose);
+		RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+		RenderSystem._setShaderTexture(0, GUI_TEXTURE);
+		blit(pose, leftPos, topPos, 0, 0, imageWidth, imageHeight);
+		super.render(pose, mouseX, mouseY, partialTicks);
+		font.draw(pose, title, width / 2 - font.width(title) / 2, topPos + 6, 4210752);
+		font.draw(pose, scrollListTitle, width / 2 - font.width(scrollListTitle) / 2, topPos + 31, 4210752);
+		ClientUtils.renderModuleInfo(pose, ModuleType.SMART, toggle, moduleRequired, isSmart, leftPos + 5, topPos + 5, width, height, mouseX, mouseY);
+	}
+
+	@Override
+	public boolean mouseDragged(double mouseX, double mouseY, int button, double deltaX, double deltaY) {
+		if (teleportationModeList != null)
+			teleportationModeList.mouseDragged(mouseX, mouseY, button, deltaX, deltaY);
+
+		return super.mouseDragged(mouseX, mouseY, button, deltaX, deltaY);
+	}
+
+	@Override
+	public boolean isPauseScreen() {
+		return false;
+	}
+
+	public class ToggleScrollList extends ScrollPanel {
+		private static final ResourceLocation BEACON_GUI = new ResourceLocation("textures/gui/container/beacon.png");
+		private final int slotHeight = 12, listLength;
+		private final List<T> orderedFilterList;
+		private final Map<T, Component> typeNames = new HashMap<>();
+
+		public ToggleScrollList(Minecraft client, int width, int height, int top, int left) {
+			super(client, width, height, top, left);
+			orderedFilterList = new ArrayList<>(be.getFilters().keySet());
+			orderedFilterList.sort((e1, e2) -> {
+				//the default entry always shows at the bottom of the list
+				if (e1 == be.getDefaultType())
+					return 1;
+				else if (e2 == be.getDefaultType())
+					return -1;
+				else
+					return Utils.localize(e1.toString()).getString().compareTo(Utils.localize(e2.toString()).getString());
+			});
+			listLength = orderedFilterList.size();
+		}
+
+		@Override
+		protected int getContentHeight() {
+			int height = listLength * (Minecraft.getInstance().font.lineHeight + 3);
+
+			if (height < bottom - top - 4)
+				height = bottom - top - 4;
+
+			return height;
+		}
+
+		@Override
+		protected boolean clickPanel(double mouseX, double mouseY, int button) {
+			int slotIndex = (int) (mouseY + (border / 2)) / slotHeight;
+
+			if (isSmart && slotIndex >= 0 && mouseY >= 0 && slotIndex < listLength) {
+				be.toggleFilter(orderedFilterList.get(slotIndex));
+				Minecraft.getInstance().getSoundManager().play(SimpleSoundInstance.forUI(SoundEvents.UI_BUTTON_CLICK, 1.0F));
+				return true;
+			}
+
+			return false;
+		}
+
+		@Override
+		public void render(PoseStack pose, int mouseX, int mouseY, float partialTick) {
+			super.render(pose, mouseX, mouseY, partialTick);
+
+			int baseY = top + border - (int) scrollDistance;
+			int mouseListY = (int) (mouseY - top + scrollDistance - (border / 2));
+			int slotIndex = mouseListY / slotHeight;
+			int slotBottom = baseY + (slotIndex + 1) * slotHeight;
+
+			if (isRedstone && mouseX >= left && mouseX <= right - 7 && slotIndex >= 0 && mouseListY >= 0 && slotIndex < listLength && mouseY >= top && mouseY <= bottom) {
+				int comparatorOutput = be.getComparatorOutputFunction().applyAsInt(orderedFilterList.get(slotIndex));
+
+				if (comparatorOutput > 0)
+					renderTooltip(pose, new TranslatableComponent("gui.securitycraft:toggleList.comparatorOutput", comparatorOutput), right - 8, slotBottom);
+			}
+		}
+
+		@Override
+		protected void drawPanel(PoseStack pose, int entryRight, int relativeY, Tesselator tess, int mouseX, int mouseY) {
+			Font font = Minecraft.getInstance().font;
+			int baseY = top + border - (int) scrollDistance;
+			int slotBuffer = slotHeight - 4;
+			int mouseListY = (int) (mouseY - top + scrollDistance - (border / 2));
+			int slotIndex = mouseListY / slotHeight;
+
+			//highlight hovered slot
+			if (isSmart && mouseX >= left && mouseX <= right - 7 && slotIndex >= 0 && mouseListY >= 0 && slotIndex < listLength && mouseY >= top && mouseY <= bottom) {
+				int min = left;
+				int max = entryRight - 6; //6 is the width of the scrollbar
+				int slotTop = baseY + slotIndex * slotHeight;
+				BufferBuilder bufferBuilder = tess.getBuilder();
+
+				RenderSystem.enableBlend();
+				RenderSystem.disableTexture();
+				RenderSystem.defaultBlendFunc();
+				bufferBuilder.begin(Mode.QUADS, DefaultVertexFormat.POSITION_COLOR);
+				bufferBuilder.vertex(min, slotTop + slotBuffer + 2, 0).uv(0, 1).color(0x80, 0x80, 0x80, 0xFF).endVertex();
+				bufferBuilder.vertex(max, slotTop + slotBuffer + 2, 0).uv(1, 1).color(0x80, 0x80, 0x80, 0xFF).endVertex();
+				bufferBuilder.vertex(max, slotTop - 2, 0).uv(1, 0).color(0x80, 0x80, 0x80, 0xFF).endVertex();
+				bufferBuilder.vertex(min, slotTop - 2, 0).uv(0, 0).color(0x80, 0x80, 0x80, 0xFF).endVertex();
+				bufferBuilder.vertex(min + 1, slotTop + slotBuffer + 1, 0).uv(0, 1).color(0x00, 0x00, 0x00, 0xFF).endVertex();
+				bufferBuilder.vertex(max - 1, slotTop + slotBuffer + 1, 0).uv(1, 1).color(0x00, 0x00, 0x00, 0xFF).endVertex();
+				bufferBuilder.vertex(max - 1, slotTop - 1, 0).uv(1, 0).color(0x00, 0x00, 0x00, 0xFF).endVertex();
+				bufferBuilder.vertex(min + 1, slotTop - 1, 0).uv(0, 0).color(0x00, 0x00, 0x00, 0xFF).endVertex();
+				bufferBuilder.end();
+				BufferUploader.end(bufferBuilder);
+				RenderSystem.enableTexture();
+				RenderSystem.disableBlend();
+			}
+
+			int i = 0;
+
+			//draw entry strings and indicators whether the filter is enabled
+			for (T type : orderedFilterList) {
+				Component name = typeNames.computeIfAbsent(type, t -> Utils.localize(t == be.getDefaultType() ? be.getDefaultTypeName() : t.toString()));
+				int yStart = relativeY + (slotHeight * i);
+
+				font.draw(pose, name, left + width / 2 - font.width(name) / 2, yStart, 0xC6C6C6);
+				RenderSystem._setShaderTexture(0, BEACON_GUI);
+				blit(pose, left, yStart - 3, 14, 14, be.getFilter(type) ? 88 : 110, 219, 21, 22, 256, 256);
+				i++;
+			}
+		}
+
+		@Override
+		public NarrationPriority narrationPriority() {
+			return NarrationPriority.NONE;
+		}
+
+		@Override
+		public void updateNarration(NarrationElementOutput narrationElementOutput) {}
+	}
+}