+++ Changelog.md
@@ -6,6 +6,7 @@
 - New: Ability to change a block pocket's outline color
 - New: The I.M.S. is now waterloggable
 - New: Scrolling in the SecurityCraft Manual while holding the control key will now scroll through subpages (if any exist)
+- New: Rift Stabilizer, a block that detects and prevents teleportation attempts in its vicinity
 - Change: The Laser Block now needs a Redstone Module in order to emit a redstone signal when someone walks through the laser
 - API: Changed LinkedAction to an interface and the enum values to records for easier usability
 - API: Several methods in LinkableBlockEntity were changed to account for the LinkedAction change